1402-05-23, c1

+++ lib/views/pages/levels/sub_bucket_page.dart
@@ -3,43 +3,77 @@ import 'package:flutter/material.dart';
 import 'package:iris_tools/api/duration/durationFormatter.dart';
 import 'package:iris_tools/modules/stateManagers/assist.dart';
 import 'package:iris_tools/widgets/irisImageView.dart';
+import 'package:pull_to_refresh/pull_to_refresh.dart';
 
-import 'package:app/pages/levels/audio_player_page.dart';
-import 'package:app/pages/levels/content_view_page.dart';
-import 'package:app/pages/levels/video_player_page.dart';
+import 'package:app/managers/mediaManager.dart';
 import 'package:app/services/favoriteService.dart';
+import 'package:app/services/lastSeenService.dart';
+import 'package:app/services/session_service.dart';
 import 'package:app/structures/abstract/stateBase.dart';
 import 'package:app/structures/enums/enums.dart';
+import 'package:app/structures/middleWares/requester.dart';
+import 'package:app/structures/models/bucketModel.dart';
 import 'package:app/structures/models/subBuketModel.dart';
 import 'package:app/system/extensions.dart';
+import 'package:app/system/keys.dart';
+import 'package:app/tools/app/appDecoration.dart';
 import 'package:app/tools/app/appDirectories.dart';
 import 'package:app/tools/app/appIcons.dart';
 import 'package:app/tools/app/appImages.dart';
 import 'package:app/tools/app/appMessages.dart';
 import 'package:app/tools/app/appThemes.dart';
 import 'package:app/tools/app/appToast.dart';
+import 'package:app/tools/app_tools.dart';
 import 'package:app/tools/routeTools.dart';
+import 'package:app/tools/searchFilterTool.dart';
 import 'package:app/views/baseComponents/appBarBuilder.dart';
+import 'package:app/views/pages/levels/audio_player_page.dart';
+import 'package:app/views/pages/levels/content_view_page.dart';
+import 'package:app/views/pages/levels/video_player_page.dart';
 import 'package:app/views/states/emptyData.dart';
+import 'package:app/views/states/errorOccur.dart';
 import 'package:app/views/states/waitToLoad.dart';
 
-class FavoritesPage extends StatefulWidget{
+class SubBucketPageInjectData {
+  BucketModel? bucketModel;
+}
+///---------------------------------------------------------------------------------
+class SubBucketPage extends StatefulWidget{
 
-  const FavoritesPage({Key? key}) : super(key: key);
+  final SubBucketPageInjectData injectData;
+
+  SubBucketPage({
+    required this.injectData,
+    Key? key,
+  }) : super(key: key);
 
   @override
-  State<FavoritesPage> createState() => _FavoritesPageState();
+  State<SubBucketPage> createState() => _SubBucketPageState();
 }
 ///==================================================================================
-class _FavoritesPageState extends StateBase<FavoritesPage> {
+class _SubBucketPageState extends StateBase<SubBucketPage> {
+  Requester requester = Requester();
+  bool isInFetchData = true;
+  String state$fetchData = 'state_fetchData';
   List<SubBucketModel> listItems = [];
+  RefreshController refreshController = RefreshController(initialRefresh: false);
+  SearchFilterTool searchFilter = SearchFilterTool();
+
 
   @override
   void initState(){
     super.initState();
 
-    assistCtr.addState(AssistController.state$loading);
-    fetchData();
+    searchFilter.limit = 20;
+    searchFilter.ascOrder = true;
+    requestData();
+  }
+
+  @override
+  void dispose(){
+    requester.dispose();
+
+    super.dispose();
   }
 
   @override
@@ -49,7 +83,7 @@ class _FavoritesPageState extends StateBase<FavoritesPage> {
         builder: (context, ctr, data) {
           return Scaffold(
             appBar: AppBarCustom(
-              title: Text(AppMessages.favorites),
+              title: Text(widget.injectData.bucketModel?.title?? ''),
             ),
             body: buildBody(),
           );
@@ -58,32 +92,56 @@ class _FavoritesPageState extends StateBase<FavoritesPage> {
   }
 
   Widget buildBody(){
-    if(assistCtr.hasState(AssistController.state$loading)) {
+    if(isInFetchData) {
       return const WaitToLoad();
     }
 
-    if(listItems.isEmpty) {
+    if(!assistCtr.hasState(state$fetchData)){
+      return ErrorOccur(onTryAgain: tryLoadClick,);
+    }
+
+    if(listItems.isEmpty){
       return const EmptyData();
     }
 
-    return GridView.builder(
-      itemCount: listItems.length,
-      gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(crossAxisCount: 2, mainAxisExtent: 200),
-      itemBuilder: (ctx, idx){
-        return buildListItem(idx);
-      },
+    return RefreshConfiguration(
+      headerBuilder: () => const MaterialClassicHeader(),
+      footerBuilder:  () => AppDecoration.classicFooter,
+      //headerTriggerDistance: 80.0,
+      //maxOverScrollExtent :100,
+      //maxUnderScrollExtent:0,
+      //springDescription: SpringDescription(stiffness: 170, damping: 16, mass: 1.9),
+      enableScrollWhenRefreshCompleted: true,
+      enableLoadingWhenFailed : true,
+      hideFooterWhenNotFull: true,
+      enableBallisticLoad: true,
+      enableLoadingWhenNoData: false,
+      child: SmartRefresher(
+        enablePullDown: false,
+        enablePullUp: true,
+        controller: refreshController,
+        onRefresh: (){},
+        onLoading: onLoadingMoreCall,
+        child: GridView.builder(
+            itemCount: listItems.length,
+            gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(crossAxisCount: 2, mainAxisExtent: 220),
+            itemBuilder: (ctx, idx){
+              return buildListItem(idx);
+            },
+        ),
+      ),
     );
   }
 
   Widget buildListItem(int idx){
     final itm = listItems[idx];
 
-    return InkWell(
-      onTap: (){
-        onItemClick(itm);
-      },
-      child: Padding(
-        padding: const EdgeInsets.all(12.0),
+    return Padding(
+      padding: const EdgeInsets.all(12.0),
+      child: InkWell(
+        onTap: (){
+          onItemClick(itm);
+        },
         child: DecoratedBox(
           decoration: BoxDecoration(
             border: Border.all(color: Colors.black26),
@@ -151,12 +209,9 @@ class _FavoritesPageState extends StateBase<FavoritesPage> {
 
                   const SizedBox(height: 12),
 
-                  Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 8.0),
-                    child: Text(itm.title, maxLines: 1).bold().fsR(1),
-                  ),
+                  Text(itm.title, maxLines: 1).bold().fsR(1),
 
-                  const SizedBox(height: 8),
+                  const SizedBox(height: 12,),
 
                   Padding(
                     padding: const EdgeInsets.symmetric(horizontal: 7.0),
@@ -181,9 +236,12 @@ class _FavoritesPageState extends StateBase<FavoritesPage> {
                             visualDensity: VisualDensity.compact,
                             iconSize: 20,
                             onPressed: (){
-                              deleteFavorite(itm);
+                              setFavorite(itm);
                             },
-                            icon: const Icon(AppIcons.delete, size: 20, color: Colors.red,)
+                            icon: Icon(itm.isFavorite ? AppIcons.heartSolid: AppIcons.heart,
+                              size: 20,
+                              color: itm.isFavorite ? Colors.red: Colors.black,
+                            )
                         )
                       ],
                     ),
@@ -197,21 +255,43 @@ class _FavoritesPageState extends StateBase<FavoritesPage> {
     );
   }
 
-  void deleteFavorite(SubBucketModel itm) async {
-    itm.isFavorite = !itm.isFavorite;
-    bool res = await FavoriteService.removeFavorite(itm.id!);
+  void tryLoadClick() async {
+    isInFetchData = true;
+    assistCtr.updateHead();
 
-    if(!res){
-      AppToast.showToast(context, AppMessages.operationFailed);
-      return;
+    requestData();
+  }
+
+  void onLoadingMoreCall(){
+    requestData();
+  }
+
+  void setFavorite(SubBucketModel itm) async {
+    itm.isFavorite = !itm.isFavorite;
+    bool res;
+
+    if(itm.isFavorite){
+      res = await FavoriteService.addFavorite(itm);
+    }
+    else {
+      res = await FavoriteService.removeFavorite(itm.id!);
     }
 
-    listItems.removeWhere((element) => element.id == itm.id);
+    if(res){
+      if(itm.isFavorite){
+        AppToast.showToast(context, AppMessages.isAddToFavorite);
+      }
+    }
+    else {
+      AppToast.showToast(context, AppMessages.operationFailed);
+    }
 
     assistCtr.updateHead();
   }
 
   void onItemClick(SubBucketModel itm) {
+    LastSeenService.addItem(itm);
+
     if(itm.type == SubBucketTypes.video.id()){
       final inject = VideoPlayerPageInjectData();
       inject.srcAddress = itm.mediaModel!.url!;
@@ -225,7 +305,7 @@ class _FavoritesPageState extends StateBase<FavoritesPage> {
       final inject = AudioPlayerPageInjectData();
       inject.srcAddress = itm.mediaModel!.url!;
       inject.audioSourceType = AudioSourceType.network;
-      inject.title = '';//widget.injectData.level1model?.title;
+      inject.title = widget.injectData.bucketModel?.title;
       inject.subTitle = itm.title;
 
       RouteTools.pushPage(context, AudioPlayerPage(injectData: inject));
@@ -241,14 +321,57 @@ class _FavoritesPageState extends StateBase<FavoritesPage> {
     }
   }
 
-  void fetchData() async {
-    listItems.addAll(FavoriteService.getAllFavorites());
+  void requestData() async {
+    final ul = AppTools.findUpperLower(listItems, searchFilter.ascOrder);
+    searchFilter.upper = ul.upperAsTS;
+    searchFilter.lower = ul.lowerAsTS;
 
-    for(final m in listItems){
-      m.isFavorite = true;
-    }
+    final js = <String, dynamic>{};
+    js[Keys.requestZone] = 'get_sub_bucket_data';
+    js[Keys.requesterId] = SessionService.getLastLoginUser()?.userId;
+    js[Keys.id] = widget.injectData.bucketModel!.id?? 1;
+    js[Keys.searchFilter] = searchFilter.toMap();
 
-    assistCtr.clearStates();
-    assistCtr.updateHead();
+    requester.bodyJson = js;
+
+    requester.httpRequestEvents.onFailState = (req, r) async {
+      isInFetchData = false;
+      assistCtr.removeStateAndUpdateHead(state$fetchData);
+    };
+
+    requester.httpRequestEvents.onStatusOk = (req, data) async {
+      isInFetchData = false;
+
+      final List bList = data['sub_bucket_list']?? [];
+      final List mList = data['media_list']?? [];
+
+      searchFilter.ascOrder = data[Keys.isAsc]?? true;
+
+      if(bList.length < searchFilter.limit){
+        refreshController.loadNoData();
+      }
+      else {
+        if(refreshController.isLoading) {
+          refreshController.loadComplete();
+        }
+      }
+
+      MediaManager.addItemsFromMap(mList);
+      MediaManager.sinkItems(MediaManager.mediaList);
+
+      for(final m in bList){
+        final itm = SubBucketModel.fromMap(m);
+        itm.isFavorite = FavoriteService.isFavorite(itm.id!);
+        itm.imageModel = MediaManager.getById(itm.coverId);
+        itm.mediaModel = MediaManager.getById(itm.mediaId);
+
+        listItems.add(itm);
+      }
+
+      assistCtr.addStateAndUpdateHead(state$fetchData);
+    };
+
+    requester.prepareUrl();
+    requester.request(context, false);
   }
 }
